rename allo adapters folder
