+++ packages/common/src/allo-adapters/adapters/allov2.QF/allocation.ts
