+++ packages/common/src/allo-adapters/hooks.ts
